Refilter the long word when taking it from cache.
Notice that node cache would store the long word when it doesn't should
to. Due to the cache is LRU based, we need to make sure the cache always
do the long word matching, otherwise we will need to have two copy of
cache. Since we have hard limit on the long word matching (3), always do
the long word matching shouldn't be costly.