Always clear matching state after cancelling.
The matching state is mostly useful for typing new character.
Typing new character will ensure the starting state is the same.
Cancelling, instead, the initial state will be different. Even if the
segment graph share the same prefix, it does not makes sense to keep the
state because the score will be different.

This fix crash if a common prefix is found during cancel().