Add naive validation on the data for history bigram. (#52)
This does not solve anything, just a simple workaround for this special
case. DATrie treat '\0' as a special value. In this case, there are 4 bytes
in the file are zero-ed, which fails DATrie internally. Ideally, we should
add a mechanism for a checksum based validation.

But add the check to "add" is not a bad thing since engine may also
wrongly pass some bad string to it. This can help skip them if such bad input
happens.