Handle customized single char final mapping
When a single char final (aeiouv) is remapped in profile, we should
consider applying the rule based on the character that it maps to, and
not apply it to the original character.

To do this, we just simply remove all stringToFinal(c) logic, and solely
relies on the finalMap_ to do the trick. finalMap_ will be filled with
extra omitted single char final entries so all the rest logic can be
unified.

Fix #62