migiate nan float issue on certain platform by introduce ...Raw() variant function to DATrie
On certain platform, nan may be canonicalized to a different value. So
DATrie<float> may need to be very careful to do comparison for check
value. So instead, we introduce Raw() function to directly return the
internal representation int32_t for value comparison. And only decode
when the value is valid.